Add git-annex & git-annex assistant
I add information on git-annex wich is a more than just a dropbox like. The git-annex assistant creates a synchronised folder on each of your OSX and Linux computers, Android devices, removable drives, and cloud services. The contents of the folder are the same everywhere. It's very easy to use, and has all the power of git and git-annex.