Update index.html
HTML is more W3C and XML compliant.

More flexible way to manage lang
A way to define the navigator langage and apply it to the page
I have not done a lot of JS, so it might not work. Moreover, It would be more optimized to search the default option langage only if it has not yet be defined with navigator.langage.