remove trailing whitespace; normalize line breaks
Having inconsistent line breaks and extra spaces makes files harder to
process automatically for anomalies or other bulk work. This cleans up
some trailing whitespace that had gotten in for a couple strings in many
languages as well as removing line breaks that are inconsistent with the
way the rest of the file was formatted.