remove Cryptocat recommendation, failed audits
The results of the last two audits (see  d2419a4) both raise significant
security and privacy concerns. Since usage is no more (and possibly
less) secure than a normal XMPP client, remove the recommendation from
PRISM-Break entirely.

This commit could easily be reversed to re-include it if/when it
receives a clean bill of health and a decision is reached to recommend
it again. That would bring back all the translations and the audit data
can be amended.