synchronize missing/unequal metadata across all locales
After this change, the only differences (I think) left are:

1. "fr-projects.json" has Wikipedia urls for the "Disconnect",
   "EtherCalc" and "Kamailio" projects, which no other locale has.

2. "zh-CN-projects.json" has translations for the "I2P" and "Ixquick"
   projects' "notes" section, which no other locale has, as introduced
   by b5495f7 from #932.

  - Notice that in the diff of b5495f7, the "notes" section for both
    these projects were empty before the change.

3. "fi-projects.json" has a translation for the "Off-the-Record
   Messaging" project's "notes" section, which no other locale has, as
   introduced by 1bd7801 from #1177.

  - Notice that in the diff of 1bd7801, the "notes" section for this
    project was empty before the change.

1 however should not be a problem because it does not look like there
are Wikipedia articles for these projects under any other language,
including English.

Note that this change does not actually check for *outdated*
translations (obviously). This change has only checked if a string type
key is empty when it should not be, or if an array type key (i.e.
"protocols" and "categories") is not equal to the rest of the keys from
all the other locales.