Adopt a Code of Conduct for project participants
PRISM-Break is largely driven by community participation in discussions
surrounding the inclusion or exclusion of various solutions. Given that
user privacy is (by definition) a very personal matter and that the
security community in general tends to have strong opinions, my
observation is that many issues that come up are fairly highly changed
even before somebody looses their cool or makes a nasty remark.

I think it would behoove the project to officially adopt a Code of
Conduct document that can be considered binding on participants in
discussion. Getting this into the repository early before any particular
major issue between participants comes up gives everybody something to
reference and a relative standard by with to judge a situation. This
could potentially save some grief vs. the possibility of working out the
details of such a code in the middle of a disputed discussion.

The Open Code of Conduct seems like a likely candidate. I'm not a huge
fan of some of the items as worded but for this project I think it will
be fine. See http://todogroup.org/opencodeofconduct/ for details.