Add Monero Currency
I have added a new cryptocurrency called Monero which unlike Bitcoin is more private and untraceable currency. Monero was coded from beginning and is not a fork of Bitcoin like any other altcoin. Here are some of the features of Monero which surpass the privacy of Bitcoin.

1) Monero's Blockchain is not public, so unlike bitcoin, you cant see the balance and transaction of a wallet, hence untraceable.
2) Monero has built-in mixers (aka Coin Mixer) which make it untraceable.
3) Monero is going to RingCT, a feature in which a transactions amount can not be sent.
4) Monero has a stealth wallet address feature in which you do not have to create multiple addresses.

So in bitcoin your privacy can be comprised by a technique called "Blockchain Analysis", which can link your transactions and might even find out where you use your money or which person withdrew it.