Simplify execution and eliminate useless variables
Having a bunch of variables that only get used once is only serving to
make the targets so hard to read nobody knows what they are doing. By
setting up the NPM binary directory in the path we can just directly
execute things from there, and by keeping commands and their
arguments all in plain view in each target it should be more apparent
what's going on and thus easier to debug and expand.