Reorganize so it's more apparent what's dependent on what
This mostly does the same thing as the previous commit with two minor
changes. In simplifying the targets with the DRY principle in mind it
became evident that there was some inconsistency in what got run when.
This actually adds building the CSS as a dependency for individual
languages as well as the whole site. At this stage this adds a little
overhead if all you are doing is tweaking language strings, but when the
Makefile is complete and basing itself on file timestamps this should
end up being a non issue as that dependency will be marked as up to date.

Note the empty-recipe terminators on some of these rules are optional
but it makes it easier to add comments and do other things with the file
without breaking the syntax is empty targets are always marked.