Re-add yarn.lock file to repository
When I commented in #1811 that having this in here didn't make sense
I was ignorant of how yarn worked. The [documentation][1] clearly
[argues][2] for why it should be tracked. On the other hand I aware of
the arguments for leaving it out, but the mostly don't apply.

Notably in the case of PRISM-Break the lock file provides some amount of
determinism to the build so that different versions of NPM modules won't
create unexpected results. This matters for stability on all projects
but it has security implications as well for our use case.

[1]: https://yarnpkg.com/lang/en/docs/yarn-lock/#toc-check-into-source-control
[2]: https://yarnpkg.com/blog/2016/11/24/lockfiles-for-all/