Drop support for npm, insist on yarn
While NMP5 would also be an option, yarn has gained quite a lot of
developer traction and I don't think it's unreasonable to settle on it
for a project like this. By sticking to one or the other we will be able
to take better advantage of features like flattening and deterministic
dependencies.