Take advantage of yarn features to secure build
By using the yarn.lock file as an intermediary we get a little bit of
self checking that we didn't have before. Note this is not strong tamper
resistance checks against the modules tree which [could be done][1] but
it is at least one step in that direction.

Additionally using yarn as a wrapper to execute binaries installed
as npm modules should mitigate path conflicts as happened in #1863.

[1]: https://stackoverflow.com/a/44226605/313192