Teach lint target to understand failure
The lint command was returning info to the console, but it wasn't
actually failing the make target so builds and even deploys could still
happen if it was failing.

This breaks it up into separate real (not phony) tasks for each input
file so that it runs _MUCH_ faster (both because it can run parallel
jobs and because it will only run for updated sources) and also so that
it actually fails the make target if there are errors.