Remove "web service" badge from Up1
Up1 is not a centralized web service, but rather a self-hosted
service. Currently there is a lot of inconsistency in usage of this
badge, but it seems to me that the only meaningful distinction is to
mark centralized web services.

Otherwise, all networking software would have to be marked as "web
service", because user will always use some particular instance of
it (think of OpenVPN and multiple OpenVPN providers).

Notably, Kolab, Nextcloud, Riot are self-hosted services that
currently do not have this tag. This pull request adds Up1 to the
list.

Specifically compare Kolab (doesn't have "web service" badge) and
Kolab Now (does have "web service" badge). I think this example
captures the essence of when this tag is applicable.