Drop Xmarks
See: https://en.wikipedia.org/wiki/Xmarks_Sync

LastPass announced on March 30, 2018 that the Xmarks service would be
shutdown on May 1, 2018. As of 11:59 UTC on May 1, 2018, the DNS entries
for the domain xmarks.com were erased and the servers were powered off
so that IP requests from web browsers that had not received the updated
DNS entries would simply time out. That is to say, the Xmarks service is
no longer operating.