Drop Server editions of proprietary operating systems
In case of Windows, Windows Server is just an edition of Windows
and as such should not be explicitly separated out.

In case of macOS, macOS Server is just an app one can install
from Mac App Store. It is as much an operating system as Nextcloud
or Wikisuite is.