Merge VPN Accounts category into VPN category under Web Services
This commit also removes CyberGhost and Hotspot Shield from "Avoid"
list because there doesn't seem to be a reason to trust those less
than all other VPN providers on the market (if there is, we should
document why and bring them back to the proprietary list). Hola!
stays because its design inherently disregards privacy and user's
traffic is sold for data scraping.