Add Borg
So far we didn't recommend any backup tools, which might lead people to
not backup their data, or just backup it unencrypted in various cloud
services.

This commit adds Borg, which is overall the best libre backup tool
that is available at the moment (read Duplicacy author's review for
reference, link below).

https://github.com/gilbertchen/duplicacy#comparison-with-other-backup-tools